[PSEC-1494] Move generic Athena logic into AwsAthenaTask
